// src/pages/createJob/changeCorporateAccount/edit/constants/types.ts
export type TFullName = {
  title: string;
  firstName: string;
  lastName: string;
};
// ...
export type TIndividualsShareholders = {
  corporateCode?: string;
  fullNames: TFullName[];
  citizenId?: string;
  passportId?: string;
  expiryDate: string;
  nationality: string;
  sharePercentage: number;
  types?: number;
  personalId?: string;
};